make flatpak (almost) buildable offline
Yeah, and it is a lot harder than it looks. Building python3-keyring
without share=network is a MASSIVE pain. Eventually I got stuck on
cryptography (look at the comment in the build-commands) and gave up
after trying forever and ever in any way I could think of of getting it
to work