initial working channel history scrolling support
This is based on how Fractal does it, in Rust, and I changed it to
python. For now autoscroll isn' useful, but can be used to improve how
the scroll-with-new-messages support works.

Also, here the performance decreases the more you scroll, as it always
fetches the full list. This should be improved in the future.

The only remaining issue I think is slightly scrolling down when the
messages finish loading. I think this is because of the spinner