entrybar: original basic redesign functionality.
the message entry bar has been redesigned to look prettier. However the
attachment and emoji buttons are currently not functional.

A possible issue that arrouse is that the message listbox border and the
entry bar should be alligned for best asthetics. They are both in a
HdyClamp, and the padding of the child is at 8px. There is a dillema
though, they must be equal, however the message listbox has to have
enough padding to look good in mobile mode, and the message entry bar
has to have a low enough amount to look round when it is in the size
group with the view switcher