add support for generic attachments
Attacments that are not recognized, like not images which are a special
case, now have a nice download button instead. And yeah, it works!

Also, I changed the attachments to need to subclass a gtk widget
themselves, not always be gtk.bin for more customizability with
composite templates.

Flatpak permission was required to xdg-download