add username role colors
We need to fetch the member not the user to get the role and from it the
color. So we achieve this by in another thread using fetch_member().

However, this causes quite a bit of rate limiting issues. This is a bit
mitigated by the random sleeping, but still presents a problem.

There is also an issue that randomly fetch_member() returns 404?

And we cannot use get_member(), for some reason it ALWAYS
returns None, even though I have the message cache enabled