add smooth scroll when attach tray is being revealed
this is done by basically setting a special variable, and then
the thing that handles the upper changing, it immediatly scrolls this to
bottom.

This has one annoying bug I wasn't able to solve. First, scroll to
bottom, then scroll back up, scroll to bottom again. And now hit the
attachment button.

The animation itself is smooth, however once the animation is complete,
a random amount of the time we might jump back down