style typing indicator
Still missing bold usernames which are a bit more complicated.
Unfortunately we needed to add a margin to the message history at the
bottom for this... Maybe with messages css upadte in the future this
will look better.

However note that the official discord client also basically does this,
however for us it is more noticable