entry_bar: send typing events when typing message
We are also expected to send this, as users do.
This would be very simple except that we need to not continually spam
discord with typing events. So we keep track of if the typing event
should already be active (trigger_typing does 10 seconds).

We also need to be an Event Receiver so that we could reset the
already_typing var when the message is sent.