add unit testing CI
pytest is added to manifest, even though it is not required for running,
because we want to run the flatpak with --command to unit test to ensure
all dependencies are met.

For the CI itself, we create a new job instead of a stage where we build
everything again because without it we get weird namespace permission
errors.

!reference is use to extend .flatpak due to the previous reason.

--non-pull and --no-related is because it takes a long time, isn't
useful, and fails due to weird separate filesystems issue.

Closes #35