confman: add signal for when a setting is changed
This is mostly useful for when a preference is changed, and an object
is meant to immediately respond to that.

This involves making confman a gobject.

I didn't create a unit test for this because it doesn't really seem
possible with pytest, as I need to wait for the callback to be called