add initial image viewer UI+boilerplate
This is the first step in creating the image viewer.

The Goal of the image viewer is to provide a convienent way to see more
of for example an image sent in the channel.

This is heavily similar to Fractal's image viewer.

The folding and animation code to work is a bit confusing due to us
wanting to put it for every channel context, instead of the main window.
So we have to deal with the Flap in a very weird way.