image_viewer: fix vertical resizing issues
First off, this was a LOT harder than it looks. Eventually I figured out
to drop the caching code from Giara, and how to calculate height by
copying Fractal.

The only minor issue left is that if scaling down based on the height,
it is slightly too much to the left