image_viewer: support navigation between media
This is basically navigation through all the found media, similar to
Fractal.

It can be a bit slow, and with a lot of overcomplication could be
significantly faster to not refetch the history every time.

The only real possible issue is that the button to navigate forward will
still appear even if you are at the very end. However checking this
would be *extremely* inneficient in the current implementation, and is
very unlikely anway. So I did not do it.