channel_sidebar: partial cleanup
The initail code was so awful, I mostly fixed it.
I didn't touch the serach code because I plan to soon redo it in a way
superior, completely different way.

Also, note about how I think instead of making the channel sidebar have
a loading spinner, havea toplevel stack to a "loading page", similar to
the one in GNOME software.