link_preview: make thread daemon
After the initial linkpreview involving tests, CI started failing,
probably because it was waiting on the thread to finiish, but still
uploading the results immediately.

Setting daemon here I think fixes that.

But the question is, what we will do in the future when we start testing
more stuff like this? Will we set daemon everywhere?

Note that this didn't impact program functionality