image_viewer: simplify and speedup catalog navigation
Turns out .history() has the convienent after= and before=parameters, so
we don't need to do this manually in a slow way.

I found out because I was fixing on how I will "jump to" in
message_view, and that I will probably need to patch discord.py again,
but I don't!