message_view: redo message adding with custom advanced functions
This is mainly to improve maintainability, and along the way fixes
message merging.

Basically, now direct changes to the listbox happen in a centralized
place.

However we do here quite a lot have to check if the object is a message
from get_children with hasattr (as I plan to be more message-like
widgets than just messages, like day separators, similar).

This also for example sets the kinetic scrolling and inserting
workarounds in a central place.

And figuring out merging was a MASSIVE pain.

First, the message now has merge and unmerge methods because I really
need to sometimes change that in runtime.

I tried to initailly have a single massive calculate merges function on
every change that is called, however the performane was AWFUL.

I ended up with a different solution:

When constructing the messages, do a simple check of if merges with the
previously getting inserted messages.

And then at the very end run an efficient fix merges function that fixes
a few edge cases (and is written to be efficient in this way).