correctly do shortcut for toggling searchbar
This is inspired after I read
https://blog.kaialexhiller.de/?p=29.

First, I now bind property of togglebutton and search-mode-enabled,
significantly easier.

And for the shortcut, now it is a PropertyAction that just changes the
search-mode-enabled property which is much easier and more convenient.

This also allows to privatize the channel sidebar and search button