channel_sidebar: redesign search
The current design was trying too much to be smart and keyboard centric
and stuff, this is a much simpler and better design that is inline with
what GNOME apps generally use.

We needed a separate search stack child because our nested expanders
don't really work well there.

I have also noticed that the channel sidebar brings quite a performance
hit (focusing/defocusing), this is probbaly because of how many
listboxes we use.

Closes #27