implement basic markdown
This uses a combination of html2pango + mistune.

The main/only current issue is that mistune creates links itself, but
html2pango also does. So I tried to make html2pango let through the
mistune links, but it seems this has created some bugs.

So I created an error handler, that if failed still only parses links
with markup_links which is 100% safe.

This is mostly OK.