add gtk4 based WebKit2 to manifest
I tested it, and it works! It took forever to build though.
Let's see how long it will take for CI :)

I also had to change sdk version to master.

I have a solution to the long build times - I created a tar archive of
the compiled version that I will use, and hopefully that will work with
Platform v40