gtk4: use binary prebuilt version of webkit
I built it first time to see it worked, and then saved the build and
uploaded it as an archive on gitlab.

Also about CI: spoiler alert: it was too long. However it was finishing
the stripping debugging symbols faze.

For Flathub I will obviously change this.

And why not type:archive? Because for some reason that extracts the
wrong directory. We can control this with manual `tar`