initial listview port
Initial basic functionaility port to GtkListView.

Loading history still weirdly has GTK-blocking performance issues,
however they were greatly improved after switching from
list_store_append to list_store_splice.

The idea with the Mobjects is that we create them from a Discord
message, or any other event. And we keep them in sync by being an
eventreceiver and sending.

I am not entirely sure how we will do non-directly-mapped-to-messages
content in the listbox, for example new message separators and the like?
Maybe the Mobject should keep that if should have a separator, and then
the widget creats it, we don't create new objects in the model for stuff
like separators.

Also not entirely sure in a similar regard how I will reimplement the
loading row/spinner.

I will also need to figure out merging messages, which I think I will be
able to do in a similar way, and maybe the mobject should have merged as
a GObject property that we can use notify on.

I temporarily disabled the Avatars due to performance issues, there is
also a question about stuff like this: should the Mobject itself lazily
load itself, not the widget?
I will try to implement binding and unbinding of the avatars to
hopefully increase performance in that regard.