message_view: reintroduce history loading indicator
With the listview this gets more complicated as we can't simply add this
completely separate widget to the start.

So instead we use a flattened listmodel with the header listmodel and
the data listmodel (as suggested on IRC). I am not sure about the
performance implications of this.

Since all objects of the listmodel must be of the same type, we add an
attribute to the mobject that signifies this is the history loading row
and disables all other functionality.

This also makes the listitem factory signal code a bit more complicated
and less clean, but I don't see a better solution to this.

I changed loading_history to a GOBject property so that I could simply
bind the spinner to it. I used props.loading_history instead of
get_property (which might be inconsitent) because it makes more sense
here