basic plugin support
This is the beginnings of a plugin engine thingy for extensibility.
A GUI will be added snoon in preferences.

I also hope to in the future to have lots of functionality be plugins,
especially stuff that is cool "wow custom client can do this" stuff
instead of actually needed functionality.

A target I am thinking of porting soon after this, is the types of login
methods (especially the GUI/graphical one).

I am not using libpeas since you need to define your own GObject
GInterfaces which isn't possible from python.

The API/usage is similar to libpeas.

Plugins are found in the plugins directory, they have a .plugin file,
gresource is supported, all the usual stuff. And plugins themselves are
subclasses of "types" of plugins.

ExtensionSets are the way to define extension points, you choose a type
of plugin you support, and you handle loading/unloading it correctly
with your type's details

I recommend looking at the documentation for more information about
this.