plugininfo: use gobject properties
This is a convenient way for property bindings and other such usecases.

And changing the `active` property should actually load/unload, however
it can't just call the load/unload methods as they themselves set the
property, so the actual loading of those functions has been split out to
a new one.

The definition of the active property is now a bit inconsistent though
with the others, as I couldn't set it with the other definition type
without weird errors.

And for the `configurable` property: this is for the GUI, and it will be
checked by checking if the get_configuration_widget method is overriden
somehow