settings_window: add plugin configuration support
As planned, however I am not really sure on passing the full preferences
window to use that function to extension row, I am not sure if that is
the recommended wayw to go about stuff like this, (And yes, I tried
doing it the other way round instead, that was even worse).

Also, maybe the extension_row should have exposed it as an action
instead? IDK

Also, there are weird harmless accesibility assertions

There are no tests because they SEGFAULT on CI, but don't locally, weird