login: port login methods to plugins
This is one of the first real uses of the plugins, I am not entirely
sure if it was worth it. But graphical login and password felt like a
"plugin" thing to me, and also, the essential manual token is marked as
"built-in".

This uses a custom plugin type: MrdLoginMethodType, and plugins are done
simply, with that they emit a token obtained signal, and they have a
content container to write to, and access to the headerbar to add extra
buttons.

The buttons and stuff are all created nicely, I think overall we now
have less duplication there>

I can't really test live creating/removing easily, as the UI isn't
accessible, and that will never be the case anyway.